fix some bugs
I added codes checking relations of initial value, minimum value, and maximum value of trait.
I fixed a bug that cause to crash when long pressing leftRange.
I fixed a bug that cause to crash when making FieldEditorActivity active, putting it in the background, leaving it for a few minutes, an bringing it to the foreground again.